requirements.txt
flask
pytest
pytest-mock
requests

# open telemetry specific
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-jaeger
opentelemetry-instrumentation-flask
